Split index.ts into multiple scripts in scripts/ and detect last known block when pulling events
